backend/requirements.txt
django==5.1.3
djangorestframework==3.15.2
requests==2.32.3
celery==5.4.0
redis==5.2.0
gunicorn==23.0.0
python-dotenv==1.0.1
pytest-django==4.9.0
drf-spectacular==0.27.2
django-redis==5.4.0
django-prometheus==2.3.1
psycopg2-binary==2.9.10
openai==1.54.3

